Cria entidade , repositório e o teste de busca por email
