+++ src/main/java/com/github/hugovallada/wallet/repository/UserRepository.java
@@ -3,5 +3,9 @@ package com.github.hugovallada.wallet.repository;
 import com.github.hugovallada.wallet.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepositoory extends JpaRepository<User, Long> {
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
 }
